Encapsulated subversion pools in the new class Pool.
* src/Makefile.in
* src/svncpp/pool.h
* src/svncpp/pool.cpp

Debugged memory hole in Client::status and Client::singleStatus.
The reason were allocations in the global pool instead of a
local sub-pool. Resolved through introduction of the new Pool class.

* src/svncpp/client.cpp (Client::status, Client::singleStatus)

Removed the remark "This class is never directly used" from the
Client class description, since it is directly used since the
introduction of the Methode status and singleStatus.

* src/svncpp/client.h



git-svn-id: http://rapidsvn.tigris.org/svn/rapidsvn/trunk@3504 65390229-12b7-0310-b90b-f21a5aa7ec8e