Update not only the folder browser but the file list as well when pressing F5 or choosing "Refresh"
* src/rapidsvn_frame.cpp
  (RapidSvnFrame::OnRefresh)

Make sure the action gets a tracer. This
was not set and since some actions are not
checking for validity of the tracer this
caused a segfault.

* src/rapidsvb_frame.cpp
  (RapidSvnFrame::Perform)

Added a virtual destructor to the add action.
But we still have a misterious error. AddAction
causes an exception when it is deleted in the
simple worker. Oh my. One of those nasty bugs.

* src/add_action.(cpp|hpp)
  (AddAction::~AddAction): added

Started to move the exception handling from
the single action to the action worker. This
makes the exception-handling and event passing
more straightforward while cleaning up the
code in the actions. The end of the action is
now to be signaled in the worker and not in the
action itself. Right now this takes place only
in AddAction. But the other actions will follow.

* src/add_action.cpp
  (AddAction::Perform): remove exception 
    handling

* src/simple_worker.(cpp|hpp)
  (SimpleWorker::PostStringEvent): added
  (SimpleWorker::PostDataEvent): added
  (SimpleWorker::Perform): send events to
    signal errors and the completion of
    and action.

Use utility function "Trace" to prevent segfaults
in the case of uninitialized m_tracer.

* src/update_action.cpp
  (UpdateAction::Perform)

Add a call to "svn_config_get_config" since the subversion
does not seem to honor a m_ctx.config = 0. Now it is
initialized.

* src/svncpp/context.cpp




git-svn-id: http://rapidsvn.tigris.org/svn/rapidsvn/trunk@5019 65390229-12b7-0310-b90b-f21a5aa7ec8e