Worked on issue #54: Ask for username/password when necessary
Action now implements ContextListener.
* src/action.(cpp|hpp)
  (Action::contextGetLogin,
   Action::contextNotify,
   Action::contextGetLogMessage): added

Enhanced logging of actions and their progress.

* src/rapidsvn_frame.cpp
  (RapidSvnFrame::OnActionEvent)
    Output log info for new id "TOKEN_ACTION_START".
    Output "Ready" after completion to the log
* src/simple_worker.cpp
  (SimpleWorker::Perfom)
    Send TOKEN_ACTION_START on start of each action
* src/ids.hpp
  (TOKEN_ACTION_START): added
* src/filelist_ctrl.cpp
  (FileListCtrl::UpdateFileList): send log status "Ready"
    at end of update

Each action now has a name that is used by the worker.
This name is set when calling Action constructor.
Removed all of the redundant exception handling code
from the actions since this will and has to be handled
in the worker. Remove the notification code at the end of
the action. This will be performded by the worker as well.
The context instance needed by some of the action is now
taken from the action using GetContext instead of creating
a local instance.

* src/action.(hpp|cpp)
  (Action::Action): new parameter name
* src/add_action.cpp
* src/copy_action.cpp
* src/merge_action.cpp
* src/delete_action.cpp
* src/checkout_action.cpp
* src/external_program_action.cpp
* src/import_action.cpp
* src/log_action.cpp
* src/simpleworker.cpp
  (SimpleWorker::Perform): create Context instance and
    pass this to the action. At the same time add this
    action as listener to the context.
* src/mkdir_action.cpp
* src/resolve_action.cpp
* src/update_action.cpp
* src/commit_action.cpp
* src/revert_action.cpp
* src/property_action.cpp
* src/cleanup_action.cpp

Move all of the private implementation data from Action
member variables to a new class Action::Data.

* src/action.(hpp|cpp)


git-svn-id: http://rapidsvn.tigris.org/svn/rapidsvn/trunk@5741 65390229-12b7-0310-b90b-f21a5aa7ec8e