Patch contributed by Tor Hovland:
Added diff support to svn::Client. Diff output is written to a
temporary file at outfile by svn_client_diff, and then the contents
of that file is read by us and returned as a string. There is also a
file for diff error messages at errfile, but I don't use it. It might
be a good idea to read that file and throw an exception if it contains
anything.

* src/svncpp/svncpp.dsp
  (SOURCE): added client_diff.cpp

* src/svncpp/Makefile.am
  (libsvncpp_la_SOURCES): added client_diff.cpp

* src/svncpp/client.hpp
  (Client::diff): added

* src/svncpp/client_diff.cpp: added
  (Client::diff): added
  (fail, diffCleanup): added




git-svn-id: http://rapidsvn.tigris.org/svn/rapidsvn/trunk@6468 65390229-12b7-0310-b90b-f21a5aa7ec8e