Issue #111: Shortcut for Commit doesnt work
Seems like this is a comparable behaviour as in #123 (No menu events with wxGTK 2.4.1).
Menu entries that have a bitmap attached do no respond to shortcuts.

We already introduced a workaround for #123, which means switching off bitmaps
for menu entries. And becaus of the annoiances in this bug the switched off bitmap entries
will be the default state. Form follows function...

* src/utils.cpp
  (UseBitmapMenus): is now false

  (CreateMenuItem): "Make directory" gets a shortcut now...

git-svn-id: http://rapidsvn.tigris.org/svn/rapidsvn/trunk@6949 65390229-12b7-0310-b90b-f21a5aa7ec8e