Issue #4: Add file diff functionality
Last changes/improvements to the diff function.

Depending on the selection there are different options available now.
If the selection is a remote directory, only two revisions can be compared
since there is no such thing like a working copy.

If the selection is a working copy then allow urls only if a single file is selected.

* src/diff_dlg.hpp src/diff_dlg.cpp
  (DiffDlg::SetDate, DiffDlg::EnableUrl, DiffDlg::AllowCompareTypes): added

* src/diff_data.hpp: renamed CompareType constants

* src/diff_action.cpp
  (DiffAction::Prepare): decide before the dialog opens what type
    of compares will be possible

  (DiffAction::Data::diffTarget): skip unversioned files and directories

  (DiffAction::Data::getFile): be more exact when logging file revision
    information (use HEAD) instead of revision 0.

git-svn-id: http://rapidsvn.tigris.org/svn/rapidsvn/trunk@6982 65390229-12b7-0310-b90b-f21a5aa7ec8e