Patch contributed by GCS:
Additional information for this patch:
The attached patch is the latest - as I found a sponsor for
RapidSVN, it was looked into and passed my mentor's approval. It will be
uploaded soon into the archive, but delayed a bit as my mentor has
problems accessing his remote repositories via RapidSVN. I believe it's
because he has way too old version, 0.32.1 actually. In this weekend he
will do some strace and gdb on it, and if it's confirmed that the
problem is not RapidSVN but the old server version, then the .deb will
be uploaded.

AND

Nope, as the docbook-xsl packages just changed, thus the configure
needs to be called with the new xsl file locations. Updated, please
commit this and not the previous one.

Sorry,
GCS

git-svn-id: http://rapidsvn.tigris.org/svn/rapidsvn/trunk@7013 65390229-12b7-0310-b90b-f21a5aa7ec8e