Issue 198: Fix folder double-clicking on Win32
Patch contributed by Frank Richter:
Folder double-clicking didn't work on Win32 as the folder to select passed to 
FolderBrowser::Data::FindClosestChild() and FolderBrowser::Data::SelectFolder()
had slashes in, but the folder in the node data has backslashes. Fixed by
copying the node path to an svn::Path instead of wxString.

* src/folder_browser.cpp
  (FolderBrowser::Data::FindClosestChild, FolderBrowser::Data::SelectFolder)

git-svn-id: http://rapidsvn.tigris.org/svn/rapidsvn/trunk@7079 65390229-12b7-0310-b90b-f21a5aa7ec8e