Issue #218: Default config dir created against user wishes
Patch contributed by Barry Scott:
Remove code that causes the default config dir to be created
when not commanded to do it.

The use of the anonymous serves no useful purpose in the library.
Infact its use in setContext will mask bugs where the caller passes
NULL when an object is the only reasonable input.

* src/svncpp/client.cpp
  (anonymous): removed
  (Client::setContext): if context is pass as null, use this and not
    anonymous.

git-svn-id: http://rapidsvn.tigris.org/svn/rapidsvn/trunk@7094 65390229-12b7-0310-b90b-f21a5aa7ec8e