Issue #230: crash when trying to expand a node in the tree view
Patch contributed by Jens Meyer auf'm Hofe:
I have not verified this on Linux, but on Windows you get a similar crash too. 
Reason is a uninitialized context in the folder browser. In revision 7094 the 
default anonymous context was removed. Because of this change a missing case 
statement in the context selection leads to this bug. See details in patch file

* src/folder_browser.cpp 
  (FolderBrowser::Data::GetContext): added switch case for FOLDER_TYPE_BOOKMARKS.

git-svn-id: http://rapidsvn.tigris.org/svn/rapidsvn/trunk@7115 65390229-12b7-0310-b90b-f21a5aa7ec8e