Issue #237: Folder Browser forgets location on refresh
Patch contributed by Jens Meyer auf'm Hofe:
SelectFolder() is called in several places, some of them with svn::Path as input 
others with wxString. While the patch above fixed the behaviour for wxString, it 
will not work with svn::Path. The solution is to convert the string to the 
subversion internal representation used in svn::Path and always use that object 
afterwards. Replace the patch above with the following one.

* src/folder_browser.cpp
  (FolderBrowser::Data::SelectFolder): Convert input string to subversion internal representation

git-svn-id: http://rapidsvn.tigris.org/svn/rapidsvn/trunk@7117 65390229-12b7-0310-b90b-f21a5aa7ec8e