Issue #221: Slow response to right click on files
Patch contributed by William Garrison:
I found the cause of this problem.  It is Windows-specific.
VerbList::InitFromDocument in verblist.cpp tries to determine if the item
right-clicked on is a file or a folder.  This function doesn't work. :-(  We
need to find another way to make this determination.

The function currently calls GetFullPath() on a filename like
"svn://MyServer/..." which becomes "\\svn\MyServer\..." Windows tries to find a
server by this name, and hangs for several seconds and gives up.

Okay, I submitted a patch for #221.  Right-clicking is now instant.  I also
disabled right-click verbs on directories since it doesn't work and doesn't make
sense.  (If you right-clicked a directory in a repository and tried to zip it or
explore it subversion would attempt that and produce an error).

* src/verblist.hpp, src/verblist.cpp
  (VerbList::InitFromDocument): added parameter "isAFolder"

* src/win32/verblist.cpp
  (VerbList::InitFromDocument): added parameter "isAFolder".
    Do not try to use wxDirExists to check whether the document is
    a file or a directory but the new parameter.

* src/utils.cpp
  (AppendVerbMenu): call InitFromDocument only for files
  
* src/external_program_action.cpp
  (ExternalProgramAction::Perform): forward m_treat_as_folder to the 
    InitFromDocument method (as parameter isAFolder)


git-svn-id: http://rapidsvn.tigris.org/svn/rapidsvn/trunk@7118 65390229-12b7-0310-b90b-f21a5aa7ec8e