Issue #267: Fixes for configure and out-of-tree builds
Patch contributed by Pavel Roskin:
The current (svn) version of RapidSVN cannot be build outside the source
tree because of an error in doc/manpage - 
rapidsvn.xml cannot be found.  The patch fixed Makefile.am to reference
the file correctly using $(srcdir).

Also, the configure script creates a file named "1" and shows error
messages it shouldn't show.  It happens because some redirections of
stderr and stdout are written incorrectly.  The patch fixes then too.
Also, it would be helpful to print the actual version of wxWidgets when
complaining about its incorrect version.

* configure.in
* doc/manpage/Makefile.am

git-svn-id: http://rapidsvn.tigris.org/svn/rapidsvn/trunk@7142 65390229-12b7-0310-b90b-f21a5aa7ec8e