Issue #288: Performance while expand item in tree view
Patch contributed by Jens Meyer auf'm Hofe (with minor
changes, especially UTF8 conversion):
While expanding a directory item in tree view it can take several seconds only 
to display a few subdirectories. The reason lies in Client::singleStatus, which 
is called for every entry in the directory. Within it the subversion function 
svn_client_status is used. For directories this function does not only get a 
single status, but it gets the status for all entries within that directory. 
Thus RapidSVN gets a lot more information than needed only to discard the rest. 
I have implemented a patch where Client::status is called only once.

There are some points to think of with this patch.
Since the information for directories comes from subversion now and not from 
filesystem, unversioned entries wont be displayed in tree view anymore. You 
still see them in file list though. Since you cant do anything with an 
unversioned entry in tree view, this is an improvement in my opinion.
The original source had some changes for UTF8 support. It is not clear to me, if 
and where these are needed in the patch. Somebody have to look at this.

* src/folder_browser.cpp
  (FolderBrowser::Data::RefreshLocal): use client.status instead of client.singleStatus
    and convert the result from utf8 to local encoding

git-svn-id: http://rapidsvn.tigris.org/svn/rapidsvn/trunk@7181 65390229-12b7-0310-b90b-f21a5aa7ec8e