Issue #358: Invalid shell syntax in configure.in
Applied patch contributed by Cory Omand:

"There are several tests which have incorrect shell syntax in configure.in.  The
one which causes the most issues is the test for --with-manpage=yes, which has
the following shell test:

if test "$MANPAGE"=="no" ; then

This is invalid syntax, and causes the man pages to never be built, even when
--with-manpage=yes is included.  I have verified that the above will evaluate to
true even if MANPAGE=yes.   The attached patch fixes this and several other tests."

Patch applied to "configure.in" partly, leaving CPPFLAGS and CXXFLAGS without changes.

* configure.in: modified

git-svn-id: http://rapidsvn.tigris.org/svn/rapidsvn/trunk@7367 65390229-12b7-0310-b90b-f21a5aa7ec8e