Issue #387: Cannot create a directory directly in repository
Called Client::mkdir() with an empty second parameter 'message', it had to be log message. But Subversion needs log message not be empty, so fails creating a directory directly in repository.

So modified Client::mkdir() having message parameter optional (for compatibility with previous releases), and the message is not passed to context, but has to be get by svn callback if necessary.

* include/svncpp/client.hpp
  (Client::mkdir): giving default value of "" to second parameter

* src/svncpp/client_modify.cpp
  (Client::mkdir): do not pass message to context, this forces svn to use callbacks
   if necessary

* src/mkdir_action.cpp
  (MkdirAction::Perform): do not specify the second parameter for client.modify()

git-svn-id: http://rapidsvn.tigris.org/svn/rapidsvn/trunk@7420 65390229-12b7-0310-b90b-f21a5aa7ec8e