Issue 502: Pass ritcher data in application
Major code-rework. Replaced all the GetBaseFlags Stuff in Actions with the
new StatusSel stuff. FolderBrowser now has a simpler implementation, that
got rid of "RefreshLocal" and "RefreshRepository" (now we have "Refresh").
Moved action creation to ActionFactory.

Now we have some unhandled exceptions after an action is finished; this
will be the next beast to track down.

git-svn-id: http://rapidsvn.tigris.org/svn/rapidsvn/trunk@7668 65390229-12b7-0310-b90b-f21a5aa7ec8e