Issue 536: Add a non-const getContext
Patch contributed by Andreas Pakulat:
A non-const getContext() method on svn::Client class would allow to create 
subclasses which use not-yet-supported features of the svn C API, like new 
version of existing functions, support more of the parameters and so on. 
Currently a const_cast<> is needed to do that.

* include/svncpp/client.hpp, src/svncpp/client.cpp
  (Client::getContext): non-const version of the method added

* AUTHORS

git-svn-id: http://rapidsvn.tigris.org/svn/rapidsvn/trunk@7694 65390229-12b7-0310-b90b-f21a5aa7ec8e