Issue 532: Add "Edit Conflicts" command (using a 3 file diff tool)
Patch contributed by Roland Rabien:
I have added a patch that implements this feature. As I am not familiar with
RapidSVN or wxWidgets, I hope I did it in a suitable way. If not, at least it
should be a good starting point for someone to do it correctly. Or just tell me
what I did wrong.

I also didn't know what to call the feature. Please let me know if I should
rename it.

* src/preferences.hpp
  (mergeTool, mergeToolArgs): added
* src/preferences.cpp: added code to save/retrieve the merge tool settings
* src/preferences_dlg.cpp: added user interface for merge tool settings
* src/action_factory.cpp: create UserResolveAction for ID_User_Resolve
* src/rapidsvn_frame.cpp: handle TOKEN_CMD_MERGE
* src/ids.hpp
* src/utils.cpp
  (AppendModifyMenu): added an entry to resolve a conflict via Merge
  (ID_User_Resolve, TOKEN_CMD_MERGE): added
* src/userresolve_action.(cpp|hpp): added

* rapidsvn.dsp, src/Makefile.am: added "userresolve_action.cpp"
* AUTHORS: added Roland Rabien to contributors

git-svn-id: http://rapidsvn.tigris.org/svn/rapidsvn/trunk@7712 65390229-12b7-0310-b90b-f21a5aa7ec8e