Patch contributed by Florian Schirmer: Hi,
please find attached to this mail a patch which fixes the instructions  
required to build RapidSVN on OSX. The resulting binary is fully  
static and a universal binary. This should fix issues 643, 645, 646  
and 647. I don't see any build script, but if you can send me a link  
to it i'm happy to fix up the nightly builds.

I had to shuffle the build order to make use of the apr-util provided  
expat library. I also upgraded the external libraries to latest stable  
versions to get rid of all workarounds / build problems. Tested on OSX  
10.5.6.

Best,
    Florian

git-svn-id: http://rapidsvn.tigris.org/svn/rapidsvn/trunk@7919 65390229-12b7-0310-b90b-f21a5aa7ec8e