Patch contributed by Florian Schirmer: please find attached a small patch which fixes the sorting speed on OSX (20x speed up). If you don't have a selected entry or switching between filter modes which will hide your current selection the focus element my be -1. This will lead to all sorts of trouble (index gets incredible huge due to signed -> unsigned casts etc).
The patch only fixes the symptoms. There are still lots of assertions  
in the list re-scrolling hack right above this code. The real root  
cause is that the indexes of topitem and focusitem are not stable  
across list refreshes. Instead of storing the index into the file list  
we're better of to store the (full) path and after rebuilding the list  
try to find the items by path?

git-svn-id: http://rapidsvn.tigris.org/svn/rapidsvn/trunk@7929 65390229-12b7-0310-b90b-f21a5aa7ec8e