Issue 661: Build instructions for Mac now without BDB (problems with static build) Patch contributed by Rachel Greenham: The build instructions in the INSTALL file for Mac OS X are out of date and do not work with 0.10 against Subversion 1.6+.
In point of fact, the main problem is with building Subversion itself with BDB
support. --with-berkeley-db=PATH enables APR to find it but the configure script
later fails with an "Invalid syntax" error; supplying --with-berkely-db in a
format as described in subversion's own INSTALL file causes APR to not even find
BDB in the first place. Catch 22. :-)

As the current 0.10 release for Macs (however its built) is PPC only, and I had
been having problems with earlier such builds, I decided to sort out an Intel
build for myself. The result is some changes to the 0.10 INSTALL file which, if
simply followed at least on an Intel/Leopard system (the most common?) results
in a successful RapidSVN.app build which even appears to work. :-)

The diff from the 0.10 INSTALL file is attached (in a moment), as is the
resulting .dmg of a Universal build of RapidSVN.app built against Subversion
1.6.2. I'd be interested to know if they do work on Panther/Tiger and with PPC,
as I can only test Intel/Leopard. But this is probably by far the most common case.

For my own purposes I earlier built a version with only Intel support (ie: not
Universal) with minimum macosx version of 10.5 and against Subversion 1.6.3.
That too worked fine, but I brought it back to Universal/OSX 10.3+/svn1.6.2 as
per existing expectations.

The main caveat, of course, is that BDB support is not enabled. But I believe
for the vastly most common usage scenario (Intel Mac running Leopard, using FSFS
for client checkouts), this is more appropriate than what's currently available. :-)

git-svn-id: http://rapidsvn.tigris.org/svn/rapidsvn/trunk@8018 65390229-12b7-0310-b90b-f21a5aa7ec8e