Issue 690: No actions for versioned entries possible for outdated entries
This problem is an side-effect of the workaround for issue 324 where we didnt get entries if non-recursive refresh-with-update as enabled. With newer version of Subversion (tested with 1.6.5) this workaround doesnt seem to be necessary anymore. So all workaround code was moved in "#ifdef WORKAROUND_ISSUE_324" which was disable which makes future removal easier.

git-svn-id: http://rapidsvn.tigris.org/svn/rapidsvn/trunk@8095 65390229-12b7-0310-b90b-f21a5aa7ec8e