Implement structured log-message control (based on wxListView).
The log message control now contains items with a specified item type
and a separate action string, in addition to the message string.
This allows to improve the readability. 
Filtering on the item types is possible by means of a new tool bar to
the left of the log message control. 
For instance, the user may desire to review only the added and
conflicted files after a large working copy update.